[UPDATE] #BTS-101 : NovelCards 목록 Response data 변경

+++ src/main/java/com/readme/sections/model/NovelCards.java
@@ -33,7 +33,7 @@ public class NovelCards {
     String serializationStatus;
     List<Tag> tags;
     Long scheduleId;
-    Float starRating;
+    Double starRating;
     Boolean monday;
     Boolean tuesday;
     Boolean wednesday;
